feat: update loading
